+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### 1.2.0: March 16th, 2022
+* Add `--no-db` and `--no-assets` flags
+* Pass `--default-character-set=utf8mb4` to `wp db export`
+
 ### 1.1.0: February 21st, 2019
 * Support for local development without a VM (Valet, etc). by passing `--local` at the end of the arguments